kata-agent/pyproject.toml
[project]
name = "kata-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=1.0.0",
    "fastapi>=0.136.0",
    "google-adk>=1.31.1",
    "google-cloud-aiplatform>=1.148.1",

    "google-cloud-documentai>=3.14.0",
    "google-cloud-firestore>=2.27.0",
    "google-cloud-storage>=3.10.1",
    "google-genai>=1.73.1",
    "httpx>=0.28.1",
    "langgraph>=1.1.8",
    "numpy>=2.4.4",
    "opencv-python-headless>=4.13.0.92",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-exporter-gcp-trace>=1.11.0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "opentelemetry-instrumentation-requests>=0.59b0",
    "opentelemetry-sdk>=1.38.0",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "pydantic-settings>=2.14.0",
    "pymupdf>=1.27.2.2",
    "python-multipart>=0.0.26",
    "uvicorn>=0.44.0",
]
